+++ python/dask_cudf/dask_cudf/tests/test_sort.py
@@ -83,7 +83,8 @@ def test_sort_repartition():
 )
 def test_sort_values_with_nulls(data, by, ascending, na_position):
     np.random.seed(0)
-    cp.random.seed(0)
+    # TODO: The uint32() exists only to support CuPy 13.2+NumPy 2
+    cp.random.seed(np.uint32(0))
     df = cudf.DataFrame(data)
     ddf = dd.from_pandas(df, npartitions=5)
